requirements.txt
python-telegram-bot==20.3
mcstatus==10.0.3
